Moving getParent into get module
