+++ lib/rules/no-timers.js
@@ -1,4 +1,4 @@
-const { getParent } = require('../utils/common');
+const { getParent } = require('../utils/get');
 const { isGuarded } = require('../utils/fast-boot');
 
 const TIMERS = ['setTimeout', 'setInterval'];
