--- lib/utils/common.js
@@ -1,16 +1,0 @@
-
-function getParent(node, predicate) {
-  while (node) {
-    if (predicate(node)) {
-      return node;
-    }
-
-    node = node.parent;
-  }
-
-  return null;
-}
-
-module.exports = {
-  getParent
-};
