+++ lib/utils/get.js
@@ -1,4 +1,4 @@
-module.exports = function get(node, path) {
+function get(node, path) {
   let parts;
   if (typeof path === 'string') {
     parts = path.split('.');
@@ -19,3 +19,20 @@ module.exports = function get(node, path) {
 
   return property;
 };
+
+function getParent(node, predicate) {
+  while (node) {
+    if (predicate(node)) {
+      return node;
+    }
+
+    node = node.parent;
+  }
+
+  return null;
+}
+
+module.exports = {
+  get,
+  getParent
+}
